use templete directly
