--- src/ui/saved-payment-methods/loading-body.jsx
@@ -1,66 +1,0 @@
-/* @flow */
-/** @jsx node */
-
-import { node, Fragment, type ChildType } from "@krakenjs/jsx-pragmatic/src";
-
-import { getStyleConfig, getStyles } from "../../zoid/saved-payment-methods/util";
-
-import { PPRebrandLogo } from "@paypal/sdk-logos/src";
-
-type SavedPaymentMethodsLoadingBodyProps = {|
-  nonce: ?string,
-  style?: Object,
-|};
-
-/**
- * Loading / prerender markup (styles + tag skeleton) shared by zoid prerender,
- * server SSR in smartcomponentnodeweb, and must match the first client paint.
- * default messages merchant config.
- */
-export function SavedPaymentMethodsLoadingBody({
-  nonce,
-  style,
-}: SavedPaymentMethodsLoadingBodyProps): ChildType {
-  const styleConfig = getStyleConfig(style);
-  const styles = getStyles(styleConfig);
-  const shouldRenderMessageContainer = Boolean(styleConfig.layout.message);
-
-  return (
-    <Fragment>
-      <style nonce={nonce}>
-        {`
-            body {
-              margin: 0;
-            }
-            ${styles}
-            `}
-      </style>
-      <div class="saved-payment-methods-container">
-        <div class="saved-payment-methods-content">
-          <div class="saved-payment-methods-container">
-            {styleConfig.layout.logo && (
-              <div class="saved-payment-methods-logo">
-                <div class="saved-payment-methods-logo-frame">
-                  <PPRebrandLogo />
-                </div>
-              </div>
-            )}
-            {styleConfig.layout.label && (
-              <div class="saved-payment-methods-label">PayPal</div>
-            )}
-            <div id="saved-payment-methods-app-root">
-              <div class="saved-payment-methods-tag-loading" />
-            </div>
-          </div>
-          {shouldRenderMessageContainer && (
-            <div
-              id="saved-payment-methods-message-root"
-              class="saved-payment-methods-message"
-              aria-hidden="true"
-            />
-          )}
-        </div>
-      </div>
-    </Fragment>
-  );
-}
