+++ src/ui/saved-payment-methods/template.jsx
@@ -1,10 +1,15 @@
 /* @flow */
 /** @jsx node */
 
-import { node, type ElementNode } from "@krakenjs/jsx-pragmatic/src";
+import { node, Fragment, type ElementNode } from "@krakenjs/jsx-pragmatic/src";
 
-import { SavedPaymentMethodsLoadingBody } from "./loading-body";
-import { validateSavedPaymentMethodsStyle } from "../../zoid/saved-payment-methods/util";
+import {
+  getStyleConfig,
+  getStyles,
+  validateSavedPaymentMethodsStyle,
+} from "../../zoid/saved-payment-methods/util";
+
+import { PPRebrandLogo } from "@paypal/sdk-logos/src";
 
 export type SavedPaymentMethodsServerRenderProps = {|
   nonce?: string,
@@ -17,13 +22,52 @@ export function validateSavedPaymentMethodsProps(props: {| style?: Object |}) {
 
 /**
  * Server-only render (commonjs2 / web: false) — mirrors smart buttons `Buttons()`.
- * Non-interactive loading shell until the iframe client app hydrates.
+ * Non-interactive loading shell (styles + tag skeleton) until the iframe hydrates;
+ * shared by zoid prerender and server SSR and must match the first client paint.
  */
 export function SavedPaymentMethods(
   props: SavedPaymentMethodsServerRenderProps
 ): ElementNode {
-  return SavedPaymentMethodsLoadingBody({
-    nonce: props.nonce,
-    style: props.style,
-  });
+  const styleConfig = getStyleConfig(props.style);
+  const styles = getStyles(styleConfig);
+  const shouldRenderMessageContainer = Boolean(styleConfig.layout.message);
+
+  return (
+    <Fragment>
+      <style nonce={props.nonce}>
+        {`
+            body {
+              margin: 0;
+            }
+            ${styles}
+            `}
+      </style>
+      <div class="saved-payment-methods-container">
+        <div class="saved-payment-methods-content">
+          <div class="saved-payment-methods-container">
+            {styleConfig.layout.logo && (
+              <div class="saved-payment-methods-logo">
+                <div class="saved-payment-methods-logo-frame">
+                  <PPRebrandLogo />
+                </div>
+              </div>
+            )}
+            {styleConfig.layout.label && (
+              <div class="saved-payment-methods-label">PayPal</div>
+            )}
+            <div id="saved-payment-methods-app-root">
+              <div class="saved-payment-methods-tag-loading" />
+            </div>
+          </div>
+          {shouldRenderMessageContainer && (
+            <div
+              id="saved-payment-methods-message-root"
+              class="saved-payment-methods-message"
+              aria-hidden="true"
+            />
+          )}
+        </div>
+      </div>
+    </Fragment>
+  );
 }
